[CDEC-401] Relocate `Bi` instances from `Pos.Binary.Ssc.Types`
