+++ ssc/src/Pos/Ssc/VssCertData.hs
@@ -75,6 +75,28 @@ flip makeLensesFor ''VssCertData
   , ("expiredCerts", "_expiredCerts")
   ]
 
+instance Bi VssCertData where
+    encode VssCertData {..} = mconcat
+        [ encodeListLen 6
+        , encode lastKnownEoS
+        -- It may look weird to encode 'VssCertificatesMap' as a
+        -- map, but it's done this way for historical reasons.
+        , encode (getVssCertificatesMap certs)
+        , encode whenInsMap
+        , encode whenInsSet
+        , encode whenExpire
+        , encode expiredCerts
+        ]
+    decode = do
+        enforceSize "VssCertData" 6
+        lastKnownEoS <- decode
+        certs <- UnsafeVssCertificatesMap <$> decode
+        whenInsMap <- decode
+        whenInsSet <- decode
+        whenExpire <- decode
+        expiredCerts <- decode
+        return VssCertData {..}
+
 -- | Create empty 'VssCertData'.
 empty :: VssCertData
 empty = VssCertData (EpochOrSlot $ Left $ EpochIndex 0) mempty mempty mempty mempty mempty
